login lecture fe for student
