+++ .env
@@ -1,1 +1,2 @@
-REACT_APP_BACKEND_URL=https://noona-todoapp-be.herokuapp.com
+# REACT_APP_BACKEND_URL=https://noona-todoapp-be.herokuapp.com
+REACT_APP_BACKEND_URL=http://localhost:5000
