+++ package.json
@@ -11,6 +11,7 @@
     "react": "^18.0.0",
     "react-bootstrap": "^2.6.0",
     "react-dom": "^18.0.0",
+    "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
