+++ src/App.css
@@ -29,6 +29,10 @@ body{
 .input-box:focus{
   outline: none;
 }
+button{
+  background-color: lightblue;
+  color: white;
+}
 .button-add{
   height: 40px;
   width: 100%;
@@ -55,4 +59,35 @@ body{
 .item-complete{
   background-color: gainsboro;
   color:white;
+}
+
+.login-box{
+  border:1px solid lightblue;
+  max-width: 600px;
+  width: 100%;
+  background-color: white;
+  padding: 20px;
+}
+.login-box label{
+  font-weight: 700;
+}
+
+.display-center{
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  
+}
+
+.button-box{
+  display:flex;
+  justify-content: space-between
+}
+
+.button-primary{
+  background-color: lightblue;
+  color: #ffffff;
+  border:none;
 }
