+++ src/pages/RegisterPage.js
@@ -1,0 +1,38 @@
+import React from "react";
+import Button from "react-bootstrap/Button";
+import Form from "react-bootstrap/Form";
+
+const RegisterPage = () => {
+  return (
+    <div className="display-center">
+      <Form className="login-box">
+        <h1>회원가입</h1>
+        <Form.Group className="mb-3" controlId="formName">
+          <Form.Label>Name</Form.Label>
+          <Form.Control type="string" placeholder="Name" />
+        </Form.Group>
+
+        <Form.Group className="mb-3" controlId="formBasicEmail">
+          <Form.Label>Email address</Form.Label>
+          <Form.Control type="email" placeholder="Enter email" />
+        </Form.Group>
+
+        <Form.Group className="mb-3" controlId="formBasicPassword">
+          <Form.Label>Password</Form.Label>
+          <Form.Control type="password" placeholder="Password" />
+        </Form.Group>
+
+        <Form.Group className="mb-3" controlId="formBasicPassword">
+          <Form.Label>re-enter the password</Form.Label>
+          <Form.Control type="password" placeholder="re-enter the password" />
+        </Form.Group>
+
+        <Button className="button-primary" type="submit">
+          회원가입
+        </Button>
+      </Form>
+    </div>
+  );
+};
+
+export default RegisterPage;
