+++ src/utils/api.js
@@ -4,6 +4,7 @@ const api = axios.create({
   baseURL: `${process.env.REACT_APP_BACKEND_URL}/api`,
   headers: {
     "Content-Type": "application/json",
+    authorization: "Bearer " + localStorage.getItem("token"),
   },
 });
 /**
